requirements.txt
discord.py>=2.0.0
aiosqlite>=0.17.0
python-dotenv>=1.0.0